R4R: uint64 bug fix

+++ modules/token/mint_token.go
@@ -6,7 +6,7 @@ import (
 
 type DocMsgMintToken struct {
 	Symbol string `bson:"symbol"`
-	Amount uint64 `bson:"amount"`
+	Amount int64  `bson:"amount"`
 	To     string `bson:"to"`
 	Owner  string `bson:"owner"`
 }
@@ -19,7 +19,7 @@ func (m *DocMsgMintToken) BuildMsg(v interface{}) {
 	msg := v.(*MsgMintToken)
 
 	m.Symbol = msg.Symbol
-	m.Amount = msg.Amount
+	m.Amount = int64(msg.Amount)
 	m.To = msg.To
 	m.Owner = msg.Owner
 }
